Rename Generics.py to ERGenerics, Generic class to ERGeneric for consistency, avoiding naming conflict with future Generic DB model
